host-tools/abtool/pyproject.toml
[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

[project]
name = "abtool"
version = "0.1.0"
description = "AnimeBoot asset pipeline: convert GIF/APNG/video to EFI-ready .anim packages."
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0,<11.0",
  "imageio>=2.34",
  "imageio-ffmpeg>=0.4",
  "numpy>=1.26"
]
readme = "USAGE.txt"

[project.scripts]
abtool = "abtool.cli:main"

[tool.setuptools.packages.find]
where = ["."]

